Fix: LLVM assertion failure with empty routes
Explanation from @asterite:
The main issue has to do with Route not being instantiated and the
compiler not knowing its instance variables. In the new compiler
approach this can't happen anymore because the compiler will be able
to figure out all types.

This patch thus instanciates a fake route, so the compiler can type
the instance variables of Route.

fixes #1