Initial take at rendering templates (ECR)
A view struct it always generated and inferred for each controller.
It is responsible for rendering the templates, with a layout.
Helpers are loaded into that struct. This is meant to avoid
polluting the controller (application logic) with the view (template
rendering).

Only ECR templates are supported, and they transformed on-the-fly to
redirect ivar accesses to the controller instead of the view class
class, to avoid pointless copying.

Being a struct, and avoiding to copy anything, we can be hopeful that
structs will be abstracted away by LLVM and be a zero-cost
abstraction.

Views are using a SafeBuffer, which means that any non-trusted value
will be HTML escaped by default. You can use `raw` on trusted
values to avoid this behavior is selected cases.

A few helpers are available, mostly related to generate safe HTML
pages. For example `#tag` and `#content_tag` are available to safely
render HTML tags. More helpers shall come!

This patch also introduces an `utils` binary (and module), meant to
access powerful features at compile time by calling to an external
binary. It includes parsing and tranforming ECR templates, listing
files in a directory (to find templates) as well as accessing
inflections (i.e. pluralize, singularize).