Overall refactor
The goal of this refactor is to simplify things and untie the different
modules, so that they stll work well together but aren't required.

Simplified `Frost::Routes` to be mostly match a route to a block, and a
simple macro to connect the route to a controller action (but you can
use plain blocks). The `namespace`, `scope` and `controller` macros are
no longer required and neither do the `resources` and `resource` that
have been removed. We don't tweak routes so often that we _need_ these
abstractions.

Simplified `Frost::View` by removing unflexible render that tied
controllers, views and the complex ECR integration (for safe output)
with simple view objects and a DSL for DOM based templates like HTML or
SVG as well as custom XML formats.

These changes helped by removing the inflector, the utils, the safe
buffers, the core type extensions, and more.

Introduces new features: `Frost::Current` and `Frost::Session`.

`Frost::Current` handles per-Fiber storage to solve accessing data from
the controller to views or to other places, such as models.

`Frost::Session` implements sessions over cookies with a memory and disk
stores.

All these features are prefectly integrated together, but none are
required per-se, not even the router or the controller, at the expense
of less magic or more explicit code (e.g. you must render views
manually, and pass data to them).